image_classification/fashion-mnist-and-mnist/requirements.txt
blinker==1.4
blis==0.4.1
boto3==1.9.4
botocore==1.12.253
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
docutils==0.15.2
Flask==1.1.2
future==0.18.2
idna==2.7
imageio==2.8.0
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
murmurhash==1.0.2
networkx==2.4
numpy==1.18.4
opencv-contrib-python==4.2.0.34
pandas==1.0.3
Pillow==7.1.2
plac==1.1.3
preshed==3.0.2
psutil==5.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.20.0
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.22.2
scipy==1.5.0
sentencepiece==0.1.91
sentry-sdk==0.15.1
six==1.14.0
spacy==2.2.4
srsly==1.0.2
thinc==7.4.0
tifffile==2020.6.3
torch==1.5.0
torchtext==0.6.0
torchvision==0.6.0
tqdm==4.46.0
urllib3==1.25.9
waitress==1.4.4
wasabi==0.7.0
Werkzeug==1.0.1
zipp==3.1.0
